consolidate insert and delete
